backend/requirements.txt
google-adk>=1.2.0
google-genai>=1.14.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=13.0
Pillow>=11.0
python-dotenv>=1.0.0
numpy>=2.0.0
aic-sdk>=2.0.0
